# docs/swagger.yaml
basePath: /v1
definitions:
  main.CreateProductPayload:
    properties:
      categories:
        items:
          type: string
        maxItems: 5
        minItems: 1
        type: array
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 100
        type: string
      price:
        type: number
    required:
    - categories
    - description
    - name
    - price
    type: object
  main.UpdateProductPayload:
    properties:
      categories:
        items:
          type: string
        maxItems: 5
        minItems: 1
        type: array
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 100
        type: string
      price:
        type: number
    type: object
  main.WishlistRequest:
    properties:
      user_id:
        type: integer
    type: object
  store.Product:
    properties:
      categories:
        items:
          type: string
        type: array
      created_at:
        description: 'TODO: Type - {Service, Item, File}'
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      reviews:
        items:
          $ref: '#/definitions/store.Review'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/store.User'
      user_id:
        type: integer
      version:
        type: integer
      wishlist:
        items:
          $ref: '#/definitions/store.UserWishlist'
        type: array
    type: object
  store.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/store.User'
      user_id:
        type: integer
    type: object
  store.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  store.UserFeedProduct:
    properties:
      categories:
        items:
          type: string
        type: array
      created_at:
        description: 'TODO: Type - {Service, Item, File}'
        type: string
      description:
        type: string
      id:
        type: integer
      is_wishlisted:
        type: boolean
      name:
        type: string
      price:
        type: number
      review_count:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/store.Review'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/store.User'
      user_id:
        type: integer
      version:
        type: integer
      wishlist:
        items:
          $ref: '#/definitions/store.UserWishlist'
        type: array
    type: object
  store.UserWishlist:
    properties:
      created_at:
        type: string
      product_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Digitally, a platform for buying and selling digital products.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Digitally API
paths:
  /healthz:
    get:
      description: Returns the current status of the API, environment, and version
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: API health check
      tags:
      - system
  /products:
    post:
      consumes:
      - application/json
      description: Creates a new product with the provided details
      parameters:
      - description: Product details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateProductPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Product'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new product
      tags:
      - products
  /products/{productID}:
    delete:
      consumes:
      - application/json
      description: Deletes a product by its ID
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retrieves a product by its ID, including its reviews
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Product'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get product by ID
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Updates a product with the provided details
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: integer
      - description: Product details to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.UpdateProductPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Product'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "409":
          description: Edit conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update product
      tags:
      - products
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Fetch the current user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetch the current user
      tags:
      - users
  /users/feed:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated feed of products for the user
      parameters:
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: desc
        description: Sort order (asc/desc)
        in: query
        name: sort
        type: string
      - description: Category to filter by
        in: query
        name: category
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Since date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: Until date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/store.UserFeedProduct'
              type: array
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get user's product feed
      tags:
      - users
  /wishlist/{productID}:
    delete:
      consumes:
      - application/json
      description: Removes a product from the user's wishlist
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: integer
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.WishlistRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Product not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Remove product from wishlist
      tags:
      - wishlist
    put:
      consumes:
      - application/json
      description: Adds a product to the user's wishlist
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: integer
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.WishlistRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Product not found
          schema: {}
        "409":
          description: Product already in wishlist
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Add product to wishlist
      tags:
      - wishlist
securityDefinitions:
  ApiKeyAuth:
    description: JWT authorization header
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"